Default argument to fetch task always set args to None
This is nonstandard behavior, I think, but that's the way it's working for me
with the latest version of paver installed.  User must now provide an args
argument when calling the paver task.  Can be an empty list.

If the default argument is left alone as an empty list, paver for some
reason thinks that it isn't being passed.  Provide a single, empty
string as an element anf paver seemsto be happy.
Branch: feature/pyinstaller