Fixing failed pyinstaller builds via versioner install params, no -p arg
Pyinstaller works best when namespace packages are installed to the same
directory tree and not installed as flat eggs (which python/pip respect
more easily).

Also, this commit removes the -p parameter passed to pyinstaller as it
was fubaring pyinstaller's ability to locate the GDAL package in the
system-site-packages.
Branch: feature/paver-updates