Allow exact matches to be given, even if match is common substring
If there are models that have multiple parts (e.g. HRA, blue carbon),
they often have a common prefix, where the main tool IS the common
prefix (so common prefix == main tool name).  We now check for this case
and run the matching tool if a match is found.
Branch: feature/invest-cli