re #BITBUCKET-3461 Looking for a "NA" value in Ratings to indicate no overlap
Katherine and Katie would like to be able to put "NA" in the ratings
column to indicate no stressor / habitat interaction.

- Add a check to see if the Ratings value is "NA"
- If "NA", set Ratings value to 0.0 so that it is deleted from dictioanry and not used in future calculations. This was already being done in the hra_preprocessor.zero_check function.
- Create a new dictionary field that indicates whether overlap should be done or not. If NA is found, then it is set to False.

Note we are saying that if the user has entered an "NA" in the Ratings column, that habitat / stressor should not be considered to have any interaction, and we handle that by saying they have no common overlap.
Branch: feature/hra_no_interaction