re #BITBUCKET-3427 adding popped argument back in for args dict.
The reason line 754 was not being reached on the test was becasue
the model pops the 'valuation_container' arg from the args dict.
So when the model is run again, the args dict has been mutated.

adding that argument back in on before the next immediate run through
Branch: feature/wave-energy-regression-tests