Correcting area change unit conversion issue
This changeset addresses a bug with the Rule-Based Scenario Generator,
where the Area Change rule was being incorrectly interpreted due to an
error in the way we were converting ha to m.  With this change, areas
are now correctly converted according to this rule.


Many thanks to Fosco Vesely for this fix.
RE #BITBUCKET-3472
Branch: feature/scenario-gen-rule-area-change-bugfix