Noting python version requirement in paver check on Windows.
On Windows, recreation requires python version 2.7.11 or higher, thanks
to a patch that was pushed to multiprocessing.py.  This patch addresses
buggy behavior in the recreation server.  Other models should work as
expected on Windows, and other OSes do not have this issue.  See the
diff of this changeset for details.

RE: #BITBUCKET-3506
Branch: bugfix/RECREATION-3506-tests-time-out-on-windows