Updating `hg paths` to `hg showconfig paths.default`.
We can't assume that a user will always have only one repo path defined,
and we can't assume that the default path will always be the last
one.  For cases where we need the default path, hg showconfig is a
much more direct way of extracting it.  RE:#BITBUCKET-3526
Branch: feature/PAVEMENT-3526-hg-showconfig