Converting model window to a MainWindow and resizing window at start.
This is an initial commit of the window resizing stuff.  It works ok,
and will expand to fill available space (but only as needed), and can be
shrunk down to a different size as needed by the user.  But at least to
start, the window will only take up as much space as needed.  RE:#BITBUCKET-3610
Branch: feature/UI-3610-natcap.ui-migration