Removing 'required' functionality from _validate().
The whole point of having required functionality in the UI itself was to
avoid a potentially expensive thread.  Threading overhead has not proven
to be an issue (anecdotally, anyways), and it's more orthogonal to have
validation handle whether an input is required.  RE:#BITBUCKET-3610  RE:#BITBUCKET-3597
Branch: feature/UI-3610-natcap.ui-migration