Renaming validation.validator to validation.invest_validator.
We can always rename this later if needed and sed the new name
throughout the codebase.  I felt that invest_validator was an
appropriate name for the decorator, and sets us up for other related
decorators down the road, should we decide to create them (e.g.
invest_execute).

I also chose not to enforce that the function name is validate.
Imposing such a restriction doesn't provide any extra benefit to us or
to the developer, and there are other places in the UI where a validate
function is expected to exist (and an error or warning is thrown if it
does not).  I'm happy to discuss this further!

RE:#BITBUCKET-3610
Branch: feature/UI-3610-natcap.ui-migration