Moving the window-reset logic to the close event handler.
The issue (#BITBUCKET-3706) was happening because the window could be closed by
pressing the X in the corner of the window, which would not trigger the
window-state-resetting.  This was only triggered when the 'Back' button
was pressed.  This is now triggered any time the window is closed, which
happens when the back button is pressed or the X is pressed.
Branch: bugfix/UI-3706-model-run-dialog-does-not-reset-between-runs